+++ projects/ixocreate/ngx-admin/src/lib/forms/types/youtube.ts
@@ -54,7 +54,6 @@ export class FormlyFieldYouTubeComponent extends CustomFieldTypeAbstract impleme
 
   setValue(value: any) {
     super.setValue(value);
-    console.log(value);
     if (this.value) {
       this.inputValue = this.youtubeUrl + this.value;
     }