+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.8...develop)
+## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.9...develop)
+
+## [v0.1.9 (2019-07-25)](https://github.com/ixocreate/admin-frontend/compare/0.1.8...0.1.9)
+### Changed
+- Reactive CustomFieldType by listening to FormControl model changes
+- Allow to set select options dynamically
+- Ixo button light color typo
 
 ## [v0.1.8 (2019-07-22)](https://github.com/ixocreate/admin-frontend/compare/0.1.7...0.1.8)
 ### Changed